setSubscription: cleanup
