+++ __tests-pacts__/serlo.org-database-layer/subscription-set-mutation.ts
@@ -46,7 +46,7 @@ test('SubscriptionSetMutation', async () => {
 
   await assertSuccessfulGraphQLMutation({
     mutation: gql`
-      mutation uuid($input: SubscriptionSetInput!) {
+      mutation setSubscription($input: SubscriptionSetInput!) {
         subscription {
           set(input: $input) {
             success
